+++ src/lib/cdm/typescript/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @dotdm/cdm
 
+## 0.6.15
+
+### Patch Changes
+
+- Updated dependencies [01d23ce]
+  - @dotdm/env@2.0.2
+  - @dotdm/contracts@3.1.2
+
+## 0.6.14
+
+### Patch Changes
+
+- Updated dependencies [002400b]
+  - @dotdm/env@2.0.1
+  - @dotdm/contracts@3.1.1
+
 ## 0.6.13
 
 ### Patch Changes